pyproject.toml
[project]
name = "agent-skills-demo"
version = "0.1.0"
description = "Minimal example of a Databricks agent with Claude Skills using LangGraph"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "databricks-langchain>=0.15.0",
    "langchain-core>=1.2.11",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "defusedxml>=0.7.1",
    "lxml>=5.0.0",
    "python-docx>=1.1.0",
    "databricks-sdk>=0.20.0",
    # For agent deployment
    "mlflow[genai]>=3.6.0",
    "databricks-agents>=1.1.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
]

[project.scripts]
agent-demo = "main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
